requirements.txt
altair==4.1.0
anyio==3.3.4
argon2-cffi==21.1.0
asgiref==3.4.1
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
black==21.9b0
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
dgl-cu111==0.7.1
einops==0.2.0
entrypoints==0.3
fastapi==0.70.0
filelock==3.3.1
gdown==3.11.0
gitdb==4.0.8
GitPython==3.1.24
h11==0.12.0
httptools==0.2.0
idna==3.3
imageio==2.9.0
imgaug==0.4.0
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lmdb==1.0.0
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.5
numpy==1.21.3
opencv-python==4.5.3.56
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prefetch-generator==1.0.1
prettytable==2.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.21
protobuf==3.19.0
ptyprocess==0.7.0
pyarrow==5.0.0
pyclipper==1.1.0.post3
pycparser==2.20
pydantic==1.8.2
pydeck==0.7.0
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.1
python-multipart==0.0.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
regex==2021.10.23
requests==2.26.0
scikit-image==0.18.3
scipy==1.7.1
Send2Trash==1.8.0
Shapely==1.7.1
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
starlette==0.16.0
streamlit==1.1.0
terminado==0.12.1
testpath==0.5.0
tifffile==2021.10.12
toml==0.10.2
tomli==1.2.1
toolz==0.11.1
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
typing-extensions==3.10.0.2
tzdata==2021.4
tzlocal==4.0.1
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
validators==0.18.2
vietocr==0.3.6
watchdog==2.1.6
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.0
widgetsnbextension==3.5.1
yapf==0.31.0

paddlepaddle=2.3.1
paddleocr=2.6
onnx=1.11.0
onnxruntime=1.10.0
streamlit=1.10.0
fastapi=0.68.1
opencv-contrib-python=4.6.0.66
opencv-python=4.6.0.66
numpy=1.19.5

